[Refactor/#6] 패키지 분리
