+++ app/src/main/AndroidManifest.xml
@@ -14,9 +14,9 @@
         android:usesCleartextTraffic="true"
         android:theme="@style/Theme.GOSOPTAndroid"
         tools:targetApi="31">
-        <activity android:name=".MainActivity"/>
-        <activity android:name=".SignUpActivity"/>
-        <activity android:name=".LoginActivity"
+        <activity android:name=".presentation.view.activity.MainActivity"/>
+        <activity android:name=".presentation.view.activity.SignUpActivity"/>
+        <activity android:name=".presentation.view.activity.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
