+++ app/src/main/java/org/android/go/sopt/data/model/SignupDto.kt
@@ -1,32 +1,34 @@
-package org.android.go.sopt.data
+package org.android.go.sopt.data.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestLoginDto(
+data class RequestSignUpDto(
     @SerialName("id")
     val id: String,
     @SerialName("password")
     val password: String,
+    @SerialName("name")
+    val name: String,
+    @SerialName("skill")
+    val skill: String
 )
 
 @Serializable
-data class ResponseLoginDto(
+data class ResponseSignUpDto(
     @SerialName("status")
     val status: Int,
     @SerialName("message")
     val message: String,
     @SerialName("data")
-    val data: LoginData,
+    val data: SignUpData,
 ) {
     @Serializable
-    data class LoginData(
-        @SerialName("id")
-        val id : String,
+    data class SignUpData(
         @SerialName("name")
         val name: String,
         @SerialName("skill")
-        val skill: String
+        val skill: String,
     )
 }
